requirements.txt
typer
pyyaml
pydantic
playwright
playwright-stealth
beautifulsoup4
lxml
pandas
openpyxl
nest_asyncio
numpy